+++ lib/absinthe/phase/debug.ex
@@ -8,7 +8,7 @@ defmodule Absinthe.Phase.Debug do
   @spec run(any, Keyword.t()) :: {:ok, Blueprint.t()}
   def run(input, _options \\ []) do
     if System.get_env("DEBUG") do
-      IO.inspect(input)
+      IO.inspect(input, label: :debug_blueprint_output)
     end
 
     {:ok, input}