+++ test/absinthe/execution/subscription_test.exs
@@ -505,7 +505,7 @@ defmodule Absinthe.Execution.SubscriptionTest do
     end
 
     @query """
-    subscription($id: String!) { otherUser(id: $id) { id } }
+    subscription($id: ID!) { otherUser(id: $id) { id } }
     """
     test "subscriptions with the same variables & document have the same subscription_ids" do
       assert {:ok, %{"subscribed" => doc1}} =
@@ -524,7 +524,7 @@ defmodule Absinthe.Execution.SubscriptionTest do
     end
 
     @query """
-    subscription($id: String!) { otherUser(id: $id) { id } }
+    subscription($id: ID!) { otherUser(id: $id) { id } }
     """
     test "subscriptions with different variables but same document have different subscription_ids" do
       assert {:ok, %{"subscribed" => doc1}} =
@@ -587,8 +587,10 @@ defmodule Absinthe.Execution.SubscriptionTest do
                context: %{pubsub: PubSub}
              )
 
-    assert_receive {[:absinthe, :execute, :operation, :start], _, %{id: id} = _meta, _config}
-    assert_receive {[:absinthe, :execute, :operation, :stop], _, %{id: ^id} = _meta, _config}
+    assert_receive {[:absinthe, :execute, :operation, :start], measurements, %{id: id}, _config}
+    assert System.convert_time_unit(measurements[:system_time], :native, :millisecond)
+
+    assert_receive {[:absinthe, :execute, :operation, :stop], _, %{id: ^id}, _config}
 
     Absinthe.Subscription.publish(PubSub, "foo", thing: client_id)
     assert_receive({:broadcast, msg})
@@ -599,8 +601,9 @@ defmodule Absinthe.Execution.SubscriptionTest do
              topic: topic
            } == msg
 
-    assert_receive {[:absinthe, :subscription, :publish, :start], _, %{id: id} = _meta, _config}
-    assert_receive {[:absinthe, :subscription, :publish, :stop], _, %{id: ^id} = _meta, _config}
+    # Subscription events
+    assert_receive {[:absinthe, :subscription, :publish, :start], _, %{id: id}, _config}
+    assert_receive {[:absinthe, :subscription, :publish, :stop], _, %{id: ^id}, _config}
 
     :telemetry.detach(context.test)
   end