+++ test/absinthe/integration/execution/operation_by_name_test.exs
@@ -2,7 +2,7 @@ defmodule Elixir.Absinthe.Integration.Execution.OperationByNameTest do
   use Absinthe.Case, async: true
 
   @query """
-  query ThingFoo($id: ID!) {
+  query ThingFoo($id: String!) {
     thing(id: $id) {
       name
     }
@@ -59,7 +59,7 @@ defmodule Elixir.Absinthe.Integration.Execution.OperationByNameTest do
   end
 
   @query """
-  mutation First($id: ID!, $thing: InputThing!) {
+  mutation First($id: String!, $thing: InputThing!) {
     first: updateThing(id: $id thing: $thing) {
       id
     }
@@ -93,7 +93,7 @@ defmodule Elixir.Absinthe.Integration.Execution.OperationByNameTest do
   end
 
   @query """
-  mutation First($id: ID!, $thing: InputThing!) {
+  mutation First($id: String!, $thing: InputThing!) {
     updateThing(id: $id thing: $thing) {
       id
     }