+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 For changes pre-v1.5 see the [v1.4](https://github.com/absinthe-graphql/absinthe/blob/v1.4/CHANGELOG.md) branch.
 
-## v1.5.0 (Beta)
+## 1.5.1
 
+- Bug Fix: #922
+
+## v1.5.0 (Rc)
+
+- Breaking Bug Fix: Variable types must align exactly with the argument type. Previously
+Absinthe allowed variables of different types to be used by accident as long as the data parsed.
+- Feature (Experimental): `:persistent_term` based schema backend
+- Breaking Change: `telemetry` event keys [changed](https://github.com/absinthe-graphql/absinthe/pull/901) since the beta release.
+
+## v1.5.0 (Beta)
 - Feature: SDL directives, other improvements
 - Feature: Output rendered SDL for a schema
 - Feature: Substantially lower subscription memory usage.